+++ recipes/affine.py
@@ -19,8 +19,7 @@ if __name__ == '__main__':
     from EinsteinEngine import *
 
     # Create a set of grid functions
-    gf = ThornDef("TestAffine", "Affine")
-    gf.set_derivative_stencil(3)
+    gf = ThornDef("TestAffine", "Affine", derivative_stencil_order=3)
 
     # Declare gfs
     g = gf.decl("g", [li, lj], symmetries=[(li, lj)], from_thorn="ADMBaseX")
