+++ recipes/waveeqn/waveeqn.py
@@ -39,10 +39,7 @@ flat_metric = mk_matrix([
 ])
 
 # Create a set of grid functions
-gf = ThornDef("TestEinsteinEngine", "WaveEqn")
-
-# Use a NRPy calculated stencil instead
-gf.set_derivative_stencil(5)
+gf = ThornDef("TestEinsteinEngine", "WaveEqn", derivative_stencil_order=5)
 
 # Declare gfs
 v_t = gf.decl("v_t", [], centering=Centering.VVV)
